requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
seaborn
scipy